[UPGRADE] Ogre3D to 1.11.5.

+++ src/main/c++/Components/MeshLodGenerator/src/OgreLodInputProviderMesh.cpp
@@ -26,10 +26,7 @@
  * -----------------------------------------------------------------------------
  */
 
-#include "OgreLodInputProviderMesh.h"
-#include "OgreLodData.h"
-#include "OgreMesh.h"
-#include "OgreSubMesh.h"
+#include "OgreMeshLodPrecompiledHeaders.h"
 
 namespace Ogre
 {
@@ -197,7 +194,7 @@ namespace Ogre
             vNormalBuf->unlock();
         }
     }
-    void LodInputProviderMesh::addIndexData(LodData* data, IndexData* indexData, bool useSharedVertexLookup, size_t submeshID, RenderOperation::OperationType renderOp)
+    void LodInputProviderMesh::addIndexData(LodData* data, IndexData* indexData, bool useSharedVertexLookup, ushort submeshID, RenderOperation::OperationType renderOp)
     {
         const HardwareIndexBufferSharedPtr& ibuf = indexData->indexBuffer;
         size_t isize = ibuf->getIndexSize();
